feat(swap): refactor navigation pt.1
- separates mobile nav & desktop navbar

- separates styling / component logic
