+++ src/app/components/Footer/index.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { PageWrapper } from '../PageWrapper';
 import { StyleConstants } from 'styles/StyleConstants';
 import styled from 'styled-components';
-import { Logo } from '../NavBar/Logo';
+import { NavLogo } from '../NavBar/NavLogo';
 import { StatisticsBanner } from '../StatisticsBanner';
 
 export function Footer() {
