+++ src/app/components/NavBar/NavBar.ts
@@ -1,0 +1,25 @@
+import styled from 'styled-components';
+import { StyleConstants } from 'styles/StyleConstants';
+
+export const NavBarWrapper = styled.header`
+  box-shadow: 0 1px 0 0 ${p => p.theme.borderLight};
+  height: ${StyleConstants.NAV_BAR_HEIGHT};
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  position: fixed;
+  top: 0;
+  width: 100%;
+  background-color: ${p => p.theme.background};
+  z-index: 2;
+  padding: 0 1.5rem;
+
+  @supports (backdrop-filter: blur(10px)) {
+    backdrop-filter: blur(10px);
+    background-color: ${p =>
+      p.theme.background.replace(
+        /rgba?(\(\s*\d+\s*,\s*\d+\s*,\s*\d+)(?:\s*,.+?)?\)/,
+        'rgba$1,0.75)',
+      )};
+  }
+`;
