+++ src/app/components/NavBar/NavLogo/index.tsx
@@ -1,0 +1,10 @@
+import { LogoWrapper, LogoTitle, LogoDescription } from './NavLogo';
+
+export function NavLogo() {
+  return (
+    <LogoWrapper>
+      <LogoTitle>SourSwap</LogoTitle>
+      <LogoDescription>next-generation DEX</LogoDescription>
+    </LogoWrapper>
+  );
+}
